requirements.txt
aiosmtplib==2.0.2
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-mail==1.4.1
Flask==3.0.3
Flask-Cors==4.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
passlib==1.7.4
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
PyGithub==2.3.0
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
retry==0.9.2
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
aiohttp
python-socketio